chore: add extract code utils
