+++ packages/ai-native/src/common/index.ts
@@ -35,7 +35,6 @@ export const AI_MENUBAR_CONTAINER_VIEW_ID = DESIGN_MENUBAR_CONTAINER_VIEW_ID;
 
 export const SLASH_SYMBOL = '/';
 export const AT_SIGN_SYMBOL = '@';
-export const BACK_QUOTE_3_SYMBOL = '```';
 
 export interface IChatMessageStructure {
   /**
